4.CNN/12.Siamese Network/siamese_network_project_v2/requirements.txt

torch
torchvision
opencv-python
streamlit
Pillow
